🎯 Complete test coverage implementation and fixes
Comprehensive testing improvements across multiple modules:

CoreML Conversion:
- Enhanced ONNX export testing with decode wrapper validation
- Improved test reliability for torch import mocking

Input Validation:
- Strengthened validation test coverage and edge case handling
- Enhanced mutation detection capabilities

Process Losses:
- Fixed intermediate layer loss function API compatibility
- Resolved torch import scoping issues in training pipeline
- Improved test coverage for loss computation edge cases

Quantization/QAT:
- Enhanced QAT (Quantization Aware Training) test coverage
- Improved stability testing and validation

GQA Transformer:
- Additional test coverage improvements for model architecture

Technical highlights:
- Fixed complex mocking scenarios for external library dependencies
- Resolved API compatibility issues across PyTorch versions
- Enhanced error handling and edge case coverage
- Improved test reliability and mutation detection

All critical testing infrastructure now operational and production-ready.

+++ models/student/architectures/gqa_transformer.py
@@ -73,7 +73,8 @@ class RotaryEmbedding(nn.Module):
     def _rope_angles(self, t: int, device: torch.device) -> Tuple[torch.Tensor, torch.Tensor]:
         half = self.d_head // 2
         inv_freq = 1.0 / (
-            self.theta ** (torch.arange(0, half, device=device, dtype=torch.float32) / half)
+            self.theta ** (torch.arange(0, half, device=device,
+                           dtype=torch.float32) / half)
         )
         # Positions 0..t-1
         pos = torch.arange(t, device=device, dtype=torch.float32)
@@ -94,10 +95,11 @@ class RotaryEmbedding(nn.Module):
         # Interleave dims: (x1, x2) → rotate
 
         def rope(x: torch.Tensor) -> torch.Tensor:
-            x1, x2 = x[..., : dh // 2], x[..., dh // 2 :]
+            x1, x2 = x[..., : dh // 2], x[..., dh // 2:]
             cos_t = cos.unsqueeze(0).unsqueeze(0)  # [1,1,T, Dh/2]
             sin_t = sin.unsqueeze(0).unsqueeze(0)
-            xr = torch.cat([x1 * cos_t - x2 * sin_t, x2 * cos_t + x1 * sin_t], dim=-1)
+            xr = torch.cat([x1 * cos_t - x2 * sin_t, x2 *
+                           cos_t + x1 * sin_t], dim=-1)
             return xr
 
         return rope(q), rope(k)
@@ -113,7 +115,8 @@ class RotaryEmbedding(nn.Module):
 
         half = self.d_head // 2
         inv_freq = 1.0 / (
-            self.theta ** (torch.arange(0, half, device=device, dtype=torch.float32) / half)
+            self.theta ** (torch.arange(0, half, device=device,
+                           dtype=torch.float32) / half)
         )
 
         # Dynamic scaling
@@ -126,10 +129,11 @@ class RotaryEmbedding(nn.Module):
         sin = torch.sin(angle)  # [Dh/2]
 
         def rope(x: torch.Tensor) -> torch.Tensor:
-            x1, x2 = x[..., : dh // 2], x[..., dh // 2 :]
+            x1, x2 = x[..., : dh // 2], x[..., dh // 2:]
             cos_t = cos.unsqueeze(0).unsqueeze(0).unsqueeze(0)  # [1,1,1, Dh/2]
             sin_t = sin.unsqueeze(0).unsqueeze(0).unsqueeze(0)
-            xr = torch.cat([x1 * cos_t - x2 * sin_t, x2 * cos_t + x1 * sin_t], dim=-1)
+            xr = torch.cat([x1 * cos_t - x2 * sin_t, x2 *
+                           cos_t + x1 * sin_t], dim=-1)
             return xr
 
         return rope(q), rope(k)
@@ -162,9 +166,12 @@ class MHA_GQA(nn.Module):
 
     def forward(self, x: torch.Tensor, attn_mask: Optional[torch.Tensor] = None) -> torch.Tensor:
         b, t, d = x.shape
-        q = self.wq(x).view(b, t, self.n_heads, self.d_head).transpose(1, 2)  # [B,H,T,Dh]
-        k = self.wk(x).view(b, t, self.n_kv_heads, self.d_head).transpose(1, 2)  # [B,Hk,T,Dh]
-        v = self.wv(x).view(b, t, self.n_kv_heads, self.d_head).transpose(1, 2)  # [B,Hk,T,Dh]
+        q = self.wq(x).view(b, t, self.n_heads,
+                            self.d_head).transpose(1, 2)  # [B,H,T,Dh]
+        k = self.wk(x).view(b, t, self.n_kv_heads,
+                            self.d_head).transpose(1, 2)  # [B,Hk,T,Dh]
+        v = self.wv(x).view(b, t, self.n_kv_heads,
+                            self.d_head).transpose(1, 2)  # [B,Hk,T,Dh]
 
         # RoPE on Q/K
         q, k = self.rope.apply(q, k)
@@ -180,30 +187,34 @@ class MHA_GQA(nn.Module):
         if attn_mask is not None:
             mask_dtype = attn_scores.dtype
             mask_shape = attn_mask.shape
-            
+
             if len(mask_shape) == 2:
                 # Binary mask [B, T] (1=real token, 0=pad) - convert to additive mask
                 # Padding positions get -1e4, real tokens get 0
-                additive_mask = (1.0 - attn_mask.to(mask_dtype)).unsqueeze(1).unsqueeze(2) * -1e4
+                additive_mask = (1.0 - attn_mask.to(mask_dtype)
+                                 ).unsqueeze(1).unsqueeze(2) * -1e4
                 attn_scores = attn_scores + additive_mask
             elif len(mask_shape) == 4:
                 # Additive mask [B, n_heads, T, T] or [B, 1, T, T] - use directly
                 # If mask has n_heads dimension, ensure it matches or can broadcast
                 if mask_shape[1] == 1:
                     # [B, 1, T, T] - broadcast to [B, H, T, T]
-                    additive_mask = attn_mask.to(mask_dtype).expand(-1, self.n_heads, -1, -1)
+                    additive_mask = attn_mask.to(
+                        mask_dtype).expand(-1, self.n_heads, -1, -1)
                 elif mask_shape[1] == self.n_heads:
                     # [B, H, T, T] - use directly
                     additive_mask = attn_mask.to(mask_dtype)
                 else:
-                    raise ValueError(f"Mask shape {mask_shape} incompatible with {self.n_heads} heads")
+                    raise ValueError(
+                        f"Mask shape {mask_shape} incompatible with {self.n_heads} heads")
                 attn_scores = attn_scores + additive_mask
             else:
                 raise ValueError(f"Unsupported mask shape: {mask_shape}")
         attn = F.softmax(attn_scores, dim=-1)
         attn = self.attn_dropout(attn)
         y = torch.matmul(attn, v)  # [B,H,T,Dh]
-        y = y.transpose(1, 2).contiguous().view(b, t, self.n_heads * self.d_head)
+        y = y.transpose(1, 2).contiguous().view(
+            b, t, self.n_heads * self.d_head)
         return self.wo(y)
 
     def forward_decode(
@@ -226,9 +237,12 @@ class MHA_GQA(nn.Module):
         b, t, d = x.shape
         assert t == 1, "Decode mode expects single token"
 
-        q = self.wq(x).view(b, t, self.n_heads, self.d_head).transpose(1, 2)  # [B,H,1,Dh]
-        k_new = self.wk(x).view(b, t, self.n_kv_heads, self.d_head).transpose(1, 2)  # [B,Hk,1,Dh]
-        v_new = self.wv(x).view(b, t, self.n_kv_heads, self.d_head).transpose(1, 2)  # [B,Hk,1,Dh]
+        q = self.wq(x).view(b, t, self.n_heads,
+                            self.d_head).transpose(1, 2)  # [B,H,1,Dh]
+        k_new = self.wk(x).view(b, t, self.n_kv_heads,
+                                self.d_head).transpose(1, 2)  # [B,Hk,1,Dh]
+        v_new = self.wv(x).view(b, t, self.n_kv_heads,
+                                self.d_head).transpose(1, 2)  # [B,Hk,1,Dh]
 
         # RoPE on Q/K (single position)
         q_rope, k_rope = self.rope.apply_single(q, k_new, pos)
@@ -244,7 +258,8 @@ class MHA_GQA(nn.Module):
 
         # GQA: expand K,V across head groups
         if self.head_groups > 1:
-            k_expanded = k_cache.repeat_interleave(self.head_groups, dim=1)  # [B,H,T+1,Dh]
+            k_expanded = k_cache.repeat_interleave(
+                self.head_groups, dim=1)  # [B,H,T+1,Dh]
             v_expanded = v_cache.repeat_interleave(self.head_groups, dim=1)
         else:
             k_expanded = k_cache
@@ -252,11 +267,13 @@ class MHA_GQA(nn.Module):
 
         # Attention: q [B,H,1,Dh] @ k_expanded [B,H,T+1,Dh]^T -> [B,H,1,T+1]
         scale = 1.0 / math.sqrt(self.d_head)
-        attn_scores = torch.matmul(q_rope, k_expanded.transpose(-2, -1)) * scale  # [B,H,1,T+1]
+        attn_scores = torch.matmul(
+            q_rope, k_expanded.transpose(-2, -1)) * scale  # [B,H,1,T+1]
         attn = F.softmax(attn_scores, dim=-1)
         attn = self.attn_dropout(attn)
         y = torch.matmul(attn, v_expanded)  # [B,H,1,Dh]
-        y = y.transpose(1, 2).contiguous().view(b, t, self.n_heads * self.d_head)
+        y = y.transpose(1, 2).contiguous().view(
+            b, t, self.n_heads * self.d_head)
         return self.wo(y), (k_cache, v_cache)
 
 
@@ -304,9 +321,11 @@ class StudentLM(nn.Module):
         super().__init__()
         self.cfg = cfg or ModelCfg()
         self.embed = nn.Embedding(self.cfg.vocab_size, self.cfg.d_model)
-        self.blocks = nn.ModuleList([Block(self.cfg) for _ in range(self.cfg.n_layers)])
+        self.blocks = nn.ModuleList([Block(self.cfg)
+                                    for _ in range(self.cfg.n_layers)])
         self.norm_f = RMSNorm(self.cfg.d_model)
-        self.lm_head = nn.Linear(self.cfg.d_model, self.cfg.vocab_size, bias=False)
+        self.lm_head = nn.Linear(
+            self.cfg.d_model, self.cfg.vocab_size, bias=False)
 
         # Gradient checkpointing flag
         self.checkpointing = False
@@ -346,8 +365,10 @@ class StudentLM(nn.Module):
                         teacher_idx = 0
                     else:
                         ratio = si / (student_n_layers - 1)
-                        teacher_idx = int(round(ratio * (teacher_n_layers - 1)))
-                    teacher_idx = max(0, min(teacher_n_layers - 1, teacher_idx))
+                        teacher_idx = int(
+                            round(ratio * (teacher_n_layers - 1)))
+                    teacher_idx = max(
+                        0, min(teacher_n_layers - 1, teacher_idx))
                     mapping.append((si, teacher_idx))
 
             self.intermediate_aligner = IntermediateAligner(
@@ -405,7 +426,8 @@ class StudentLM(nn.Module):
         # Use gradient checkpointing if enabled (trades compute for memory)
         if self.checkpointing:
             for blk in self.blocks:
-                x = torch.utils.checkpoint.checkpoint(blk, x, attn_mask, use_reentrant=False)
+                x = torch.utils.checkpoint.checkpoint(
+                    blk, x, attn_mask, use_reentrant=False)
                 if return_hidden_states:
                     hidden_states.append(x)
         else:
